... haiku is useless
